refactor: separate handling messages to different files
