+++ src/api/dataBase.ts
@@ -22,7 +22,7 @@ export const isUserExist = (
   }
   if (type === 'login') {
     colorConsole.red(
-      `User with name ${name} and password ${password} has already login.`,
+      `User with name "${name}" and password "${password}" has already login.`,
     );
   }
   return existingUser;
@@ -35,7 +35,7 @@ export const addNewUser = (user: IUser, type: 'login' | 'register') => {
     db.push(user);
     return true;
   }
-  colorConsole.red(`User with index ${user.index} already exists.`);
+  colorConsole.red(`User with index "${user.index}" has already exists.`);
   return false;
 };
 
@@ -44,9 +44,9 @@ export const deleteUser = (userId: number) => {
     user.index === userId;
   });
   if (userIndex !== -1) {
-    colorConsole.red(`User with index ${userId} not found.`);
+    colorConsole.red(`User with index "${userId}" is not found.`);
     return false;
   }
-  colorConsole.red(`User with index ${userId} not found.`);
-  return false;
+  loginUserDB.splice(userIndex, 1);
+  return true;
 };
